refactor: Price, quantity 필드 타입 변경

+++ src/main/java/com/example/shopping/domain/order/entity/OrderItem.java
@@ -26,6 +26,6 @@ public class OrderItem {
     @JoinColumn(name = "order_id", nullable = false)
     private Order order;
 
-    private Integer price;
-    private Integer quantity;
+    private Long price;
+    private Long quantity;
 }
